+++ CHANGELOG.md
@@ -14,8 +14,36 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
 ## [Unreleased]
+### Other changes:
+- Add backup reminder, see settings for more options
+
 ### Bugfix 🐛:
 - Fixed a bug that caused a small icon to be displayed in the F-Droid store (German).
+
+
+## [0.7.2] - 2024-09-22
+### Added Features and Improvements 🙌:
+- Hello World! Thx to the community, the app is now available in Croatian 🎉
+- Using the latest flutter 3.24 with upgraded deps
+
+### Bugfix 🐛:
+- Fix showing ukraine as supported language
+
+
+## [0.7.1] - 2024-07-03
+### Added Features and Improvements 🙌:
+- Hello World! Thx to the community, the app is now available in Turkish 🎉
+
+### Other changes:
+- Upgraded dependencies
+- Improve readability of target weight label
+
+### Bugfix 🐛:
+- Fix broken color of linechart
+- Allow adding measurements older than 2 years
+- Fix bug of showing target weight correctly using st/lb
+- Fixed a bug where saving an unmodified measurement resulted in it being deleted
+>>>>>>> main
 
 
 ## [0.7.0] - 2024-05-29
@@ -184,7 +212,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - initial release
 
 
-[Unreleased]: https://github.com/quantumphysique/trale/compare/v0.7.0...main
+[Unreleased]: https://github.com/quantumphysique/trale/compare/v0.7.2...main
+[0.7.2]: https://github.com/quantumphysique/trale/compare/v0.7.1...v0.7.2
+[0.7.1]: https://github.com/quantumphysique/trale/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/quantumphysique/trale/compare/v0.6.2...v0.7.0
 [0.6.2]: https://github.com/quantumphysique/trale/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/quantumphysique/trale/compare/v0.5.0...v0.6.1